// expo/constants/types.ts
export type MealType = 'breakfast' | 'lunch' | 'dinner' | 'snack';

export interface MealLog {
  id: string;
  date: Date;
  name: string;
  mealType?: MealType;
  calories: number;
  protein: number;
  carbs: number;
  fat: number;
  sodiumMg: number;
  fiber?: number;
  notes?: string;
  imageUri?: string;
  customFoodId?: string;
}